Should not use poderse, instead use poder
Poderse is the reflexive version of the verb poder. Poder means "to be able to" in this context. Poderse in this context would be a generalization (voz pasiva refleja). 

The goal is to make it say "Could not combine keys". "No se puede combinar claves" essentially means "Keys cannot be combined" which is a false statement. Instead, we want to say "No puede combinar claves" which means "Could not combine keys".