Tagline update (PNG needs to be re-created)
The PNG would need to be regenerated from the SVG file. I don't have permissions I think to upload the modified PNG. I uploaded a copy to http://imgur.com/wtNicXB it would just have to be renamed to peercoin_ltext.png. Just to note that there was also a logo_text_tagline.png file in the img/logos folder but it doesn't seem to be used anywhere when it searched through the page code.